+++ cpp/hello/src/callback.rs
@@ -1,11 +1,11 @@
-use libc::{c_int, size_t};
+use libc::{c_float, c_int, size_t};
 
 #[no_mangle]
 pub extern "C" fn hello__callback__map_safe(
     items_in: *const c_int,
-    items_out: *mut c_int,
+    items_out: *mut c_float,
     item_count: size_t,
-    mapper: extern "C" fn(c_int) -> c_int,
+    mapper: extern "C" fn(c_int) -> c_float,
 ) {
     let items: &[c_int] = unsafe { std::slice::from_raw_parts(items_in, item_count) };
     let mapped_items = hello::callback::map(items, |item| mapper(*item));
